requirements.txt
streamlit
pdfplumber
python-docx
spacy
scikit-learn